increased data-type size for QueryConfig limit and offset (Integer -> Long)

+++ src/main/java/org/dynapi/dynapi/core/query/QueryConfig.java
@@ -12,11 +12,11 @@ public class QueryConfig {
     /**
      * maximum amount to query (for pagination)
      */
-    private Integer limit = null;
+    private Long limit = null;
     /**
      * offset for pagination
      */
-    private Integer offset = null;
+    private Long offset = null;
     /**
      * controls the order by
      */
